smart completion of data types inside of instructions
